Enabling more modules
I don't see the need in selectively including syntax for modules when I can
just install the modules and include them all. Expanded to include all of the
modules I could that were in the list already and updated. Going to look into
getting a more up to date module list for PHP 5.5+ and expanding further.